requirements.txt
opencv-python
tensorflow
pytorch
numpy 
pandas
Pillow